requirements.txt
flask==2.0.1
firebase-admin==5.0.0
gunicorn==20.1.0
tensorflow==2.7.0
numpy==1.19.5
scikit-learn==0.24.2
joblib==1.0.1
python-dotenv==0.19.0
flask-login==0.5.0
werkzeug==2.0.1
flask-wtf==0.15.1
email-validator==1.1.3